+++ lib/routines/builtin/__ashlti3.c
@@ -1,5 +1,5 @@
-int __ashlsi3(int a, int b) {
-    int result = a;
+long long __ashlsi3(long long a, int b) {
+    long long result = a;
 
     for (int i = 0; i < b; i++)
     {
